src/main/java/edu/ucsb/cs156/spring/hello/Developer.java
package edu.ucsb.cs156.spring.hello;

/**
 * A class with static methods to provide information about the developer.
 */

public class Developer {

    // This class is not meant to be instantiated
    // so we make the constructor private

    private Developer() {}
    
    /**
     * Get the name of the developer
     */

    public static String getName() {
        

        return "Amber W.";
    }

    /**
     * Get the github id of the developer
     * @return github id of the developer
     */

    public static String getGithubId() {
        return "yw2398";
    }

    /**
     * Get the developers team
     * @return developers team as a Java object
     */
    
    public static Team getTeam() {
       
        Team team = new Team("f24-08");
        team.addMember("Amber");
        team.addMember("Caleb");
        team.addMember("Christy");
        team.addMember("Divyani");
        team.addMember("Mike");
        team.addMember("Oscar");
        return team;
    }
}
